CR.FIX.3675 removed option create passwords for SOLindexer/SDXdesigner modules X.3609 restrict output passwords from pdfutility, sdxdesigner, and solscript products Prev Revisions: CR.FIX.3933 initial checkin of spd-e key addition code Prev Revisions: CR.FIX.3933 initial checkin of spd-e key addition code added document assembler

+++ CreatorCode/KEY/NewKey/pkey.h
@@ -198,7 +198,7 @@ public:
    void getBasePassword(ISolimarLicenseSvr* pServer, char* Password_String);
    enum {MAX_EXTENSION_NUM = 4};
 
-   void getExtensionPassword(uchar extension_days,
+   void getExtensionPassword(ushort extension_days,
                               ushort extension_num,
                               ISolimarLicenseSvr* pServer,
                               char* password);
